Cleaned up files, deleted comments
