+++ src/redux/reducers/clientDashboardStoreReducer.js
@@ -1,0 +1,10 @@
+const clientDashboardStoreReducer = ( state = [], action ) => {
+    switch (action.type) {
+        case 'SET_CLIENT_DASHBOARD_STORES':
+            return action.payload;
+        default:
+            return state;
+    }
+}
+
+export default clientDashboardStoreReducer;
