+++ src/redux/reducers/index.js
@@ -7,6 +7,7 @@ import locationsReducer from './locationsReducer';
 import supportReducer from './supportReducer';
 import clientStoreReducer from './clientStoreReducer';
 import individualStoreReducer from './individualStoreReducer';
+import clientDashboardStoreReducer from './clientDashboardStoreReducer';
 
 // rootReducer is the primary reducer for our entire project
 // It bundles up all of the other reducers so our project can use them.
@@ -23,6 +24,7 @@ const rootReducer = combineReducers({
   supportReducer, // will handle support ticket interaction
   clientStoreReducer, // will handle a single client store
   individualStoreReducer,
+  clientDashboardStoreReducer,
 });
 
 export default rootReducer;
